feat(frontend): Restructure app components folder
