+++ frontend/src/lib/components/apps/components/dataDisplay/AppBarChart.svelte
