+++ frontend/src/lib/components/apps/components/dataDisplay/AppPieChart.svelte
