+++ frontend/src/lib/components/apps/components/dataDisplay/AppText.svelte
